TinLM 6/11/2018 commit 5fc9f7c2f8fc87771a8746f43403742392f5bab5. Payment

+++ app/src/main/java/vu/huy/bookhouse/model/User.java
@@ -8,13 +8,13 @@ public class User {
     private String username = "";
     private String passsword = "";
     private String fullname = "";
-    private double balance = 0;
+    private float balance = 0;
     private Date VIPEndDate;
 
     public User() {
     }
 
-    public User(String userID, String username, String passsword, String fullname, double balance, Date VIPEndDate) {
+    public User(String userID, String username, String passsword, String fullname, float balance, Date VIPEndDate) {
         this.userID = userID;
         this.username = username;
         this.passsword = passsword;
@@ -55,11 +55,11 @@ public class User {
         this.fullname = fullname;
     }
 
-    public double getBalance() {
+    public float getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(float balance) {
         this.balance = balance;
     }
